+++ CHANGELOG.md
@@ -14,5 +14,6 @@ For information on changes in released versions of Teku, see the [releases page]
 ### Breaking Changes
 
 ### Additions and Improvements
+* Added `kintsugi` network definition. 
 
 ### Bug Fixes